🐞 Sample manifests don't return 403 Forbidden
Sample manifests don't return 403 "Forbidden" or emit `Permission
denied` (ENOPERM) in the log files.

This is caused by the new line of stemcells, 3541+, which have a much
more stringent `umask`.

We fix this by setting the nginx worker to run as UNIX group `vcap`,
which has permission to view most directories/files.

[fixes #13]